Extract ESLint formatter so other modules can use it.
