+++ Lombiq.NodeJs.Extensions/scripts/lint-scripts.js
@@ -1,19 +1,8 @@
 ﻿const path = require('path');
 const { ESLint } = require("eslint");
-const { handleWarningObject, handleErrorMessage } = require('./handle-error');
 
-function formatResult(result) {
-    result.messages?.forEach((warning) =>
-        handleWarningObject({
-            message: warning.fix
-                ? `${warning.message} (An automatic fix is available with the ESLint CLI.)`
-                : warning.message,
-            code: warning.ruleId,
-            path: result.filePath,
-            line: warning.line,
-            column: warning.column,
-        }));
-}
+const formatter = require('./eslint-msbuild-formatter');
+const { handleErrorMessage } = require('./handle-error');
 
 const options = {
     cwd: path.resolve(process.argv.length > 2 ? process.argv[2] : '.'),
@@ -29,7 +18,7 @@ const options = {
     if (!Array.isArray(results) || results.length === 0) return;
 
     // 3. Format the results.
-    results.forEach(formatResult);
+    formatter(results);
 })().catch((error) => {
     handleErrorMessage(error);
     process.exit(1);
